Increase the range of extents.first objects.  At SBND resolution, max image count per file is 595.  THis update puts it about 1 trillion.

+++ src/larcv3/core/dataformat/DataFormatTypes.h
@@ -28,12 +28,12 @@ namespace larcv3 {
   static const ProjectionID_t kINVALID_PROJECTIONID = kINVALID_USHORT;
 
   struct Extents_t{
-    unsigned int first;
+    unsigned long long int first;
     unsigned int n;
   };
 
   struct IDExtents_t{
-    unsigned int first;
+    unsigned long long int first;
     unsigned int n;
     unsigned int id;
   };
